requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
streamlit
openpyxl
pyserial
streamlit-autorefresh
flask
requests